renaming prometheus to be a feature
