+++ src/app.js
@@ -1,7 +1,6 @@
 const express = require('@feathersjs/express');
 const feathers = require('@feathersjs/feathers');
 const configuration = require('@feathersjs/configuration');
-const apiMetrics = require('prometheus-api-metrics');
 const path = require('path');
 const favicon = require('serve-favicon');
 const compress = require('compression');
@@ -11,7 +10,6 @@ const bodyParser = require('body-parser');
 const socketio = require('@feathersjs/socketio');
 const { ObjectId } = require('mongoose').Types;
 
-const { Configuration } = require('@schul-cloud/commons');
 const { BODYPARSER_JSON_LIMIT, LEAD_TIME } = require('../config/globals');
 
 const middleware = require('./middleware');
@@ -24,6 +22,7 @@ const requestLogger = require('./middleware/requestLogger');
 const errorHandler = require('./middleware/errorHandler');
 const sentry = require('./middleware/sentry');
 const rabbitMq = require('./utils/rabbitmq');
+const prometheus = require('./utils/prometheus');
 
 const setupSwagger = require('./swagger');
 const { initializeRedisClient } = require('./utils/redis');
@@ -44,17 +43,7 @@ const setupApp = async () => {
 		});
 	}
 
-	if (Configuration.has('PROMETHEUS_ENABLED') && Configuration.get('PROMETHEUS_ENABLED') === true) {
-		const metricsOptions = {};
-		if (Configuration.has('PROMETHEUS__METRICS_PATH')) {
-			metricsOptions.metricsPath = Configuration.get('PROMETHEUS__METRICS_PATH');
-		}
-		if (Configuration.has('PROMETHEUS__DURATION_BUCKETS_SECONDS[0]')) {
-			// TODO rewrite configuration to support arrays via get()
-			metricsOptions.durationBuckets = Configuration.data.PROMETHEUS.DURATION_BUCKETS_SECONDS;
-		}
-		app.use(apiMetrics(metricsOptions));
-	}
+	app.configure(prometheus);
 
 	setupSwagger(app);
 	initializeRedisClient();
